organization(naming): changed component names to fit our naming conventions
