+++ libs/organization/src/index.ts
@@ -1,7 +1,7 @@
 export * from './lib/+state';
 export * from './lib/organization.module';
 export * from './lib/guard/organization.guard'
-export * from './lib/components/organization-form/organization-form.component';
+export * from './lib/components/organization-create/organization-create.component';
 export * from './lib/permissions/guard/permissions.guard'
 export * from './lib/permissions/+state/permissions.service'
 export * from './lib/pages/organization-home/organization-home.component';
