+++ libs/organization/src/lib/components/organization-create/organization-create.component.html
@@ -1,0 +1,35 @@
+<section>
+  <mat-card>
+    <mat-card-header>
+      <mat-card-title>Principal Informations</mat-card-title>
+    </mat-card-header>
+    <mat-card-content>
+      <form [formGroup]="form" (ngSubmit)="addOrganization()">
+        <h4>Organization Name</h4>
+        <mat-form-field appearance="outline">
+          <mat-label>Name</mat-label>
+          <input matInput placeholder="Name" aria-label="name" formControlName="name" />
+        </mat-form-field>
+
+        <h4>Organization Address</h4>
+        <mat-form-field appearance="outline">
+          <mat-label>Address</mat-label>
+          <textarea
+            matInput
+            placeholder="Address"
+            aria-label="address"
+            formControlName="address"
+          ></textarea>
+        </mat-form-field>
+        <mat-toolbar>
+          <mat-toolbar-row>
+            <div>
+              <button mat-button type="button" (click)="clear()">Cancel</button>
+              <button mat-raised-button type="submit" color="primary">Next</button>
+            </div>
+          </mat-toolbar-row>
+        </mat-toolbar>
+      </form>
+    </mat-card-content>
+  </mat-card>
+</section>
