+++ libs/organization/src/lib/components/organization-create/organization-create.component.scss
@@ -1,0 +1,34 @@
+:host {
+  display: block;
+  height: 100%;
+  padding: 2rem;
+  margin: 100px 300px;
+
+  .mat-form-field {
+    width: 100%;
+  }
+
+  .mat-card-title {
+    font-size: 1.8rem;
+  }
+
+  .mat-toolbar {
+    background: transparent;
+
+    .mat-toolbar-row {
+      display: flex;
+      justify-content: center;
+      align-items: center;
+
+      > div {
+        width: 55%;
+      }
+
+      button {
+        margin: 0.5rem;
+        width: 50%;
+      }
+    }
+  }
+}
+
