--- libs/organization/src/lib/components/organization-edit/organization-edit.component.html
@@ -1,23 +1,0 @@
-<h4>Picture</h4>
-<!-- TODO: issue#579 import picture -->
-<div fxLayoutAlign="center">
-  <img src="https://material.angular.io/assets/img/examples/shiba2.jpg" alt="Profile picture" />
-</div>
-<article class="actions" fxLayout="column" fxLayoutAlign="center center">
-  <button mat-raised-button color="accent">Import a picture</button>
-  <button mat-button color="primary">Delete this picture</button>
-</article>
-<mat-divider></mat-divider>
-<article>
-  <form [formGroup]="control" fxLayout="column" fxLayoutAlign="center start">
-    <h4>Informations</h4>
-    <mat-form-field appearance="outline">
-      <mat-label>Address</mat-label>
-      <input matInput placeholder="Position" type="text" formControlName="address" />
-    </mat-form-field>
-    <mat-form-field appearance="outline">
-      <mat-label>Phone number</mat-label>
-      <input matInput placeholder="Phone number" type="tel" formControlName="phoneNumber" />
-    </mat-form-field>
-  </form>
-</article>
