--- libs/organization/src/lib/components/organization-edit/organization-edit.component.scss
@@ -1,18 +1,0 @@
-mat-form-field {
-  width: 100%;
-}
-
-img {
-  width: 200px;
-  height: 200px;
-}
-
-h4 {
-  margin-bottom: 16px;
-  padding-top: 16px;
-  color: var(--lightGrey);
-}
-
-.actions {
-  padding: 8px 0;
-}
