--- libs/organization/src/lib/components/organization-edit/organization-edit.component.ts
@@ -1,16 +1,0 @@
-import { Component, ChangeDetectionStrategy } from '@angular/core';
-import { ControlContainer } from '@angular/forms';
-
-@Component({
-  selector: 'organization-edit',
-  templateUrl: './organization-edit.component.html',
-  styleUrls: ['./organization-edit.component.scss'],
-  changeDetection: ChangeDetectionStrategy.OnPush
-})
-export class OrganizationEditComponent {
-  constructor(public controlContainer: ControlContainer) {}
-
-  public get control() {
-    return this.controlContainer.control;
-  }
-}
