+++ libs/organization/src/lib/components/organization-form/organization-form.component.html
@@ -1,35 +1,23 @@
-<section>
-  <mat-card>
-    <mat-card-header>
-      <mat-card-title>Principal Informations</mat-card-title>
-    </mat-card-header>
-    <mat-card-content>
-      <form [formGroup]="form" (ngSubmit)="addOrganization()">
-        <h4>Organization Name</h4>
-        <mat-form-field appearance="outline">
-          <mat-label>Name</mat-label>
-          <input matInput placeholder="Name" aria-label="name" formControlName="name" />
-        </mat-form-field>
-
-        <h4>Organization Address</h4>
-        <mat-form-field appearance="outline">
-          <mat-label>Address</mat-label>
-          <textarea
-            matInput
-            placeholder="Address"
-            aria-label="address"
-            formControlName="address"
-          ></textarea>
-        </mat-form-field>
-        <mat-toolbar>
-          <mat-toolbar-row>
-            <div>
-              <button mat-button type="button" (click)="clear()">Cancel</button>
-              <button mat-raised-button type="submit" color="primary">Next</button>
-            </div>
-          </mat-toolbar-row>
-        </mat-toolbar>
-      </form>
-    </mat-card-content>
-  </mat-card>
-</section>
+<h4>Picture</h4>
+<!-- TODO: issue#579 import picture -->
+<div fxLayoutAlign="center">
+  <img src="https://material.angular.io/assets/img/examples/shiba2.jpg" alt="Profile picture" />
+</div>
+<article class="actions" fxLayout="column" fxLayoutAlign="center center">
+  <button mat-raised-button color="accent">Import a picture</button>
+  <button mat-button color="primary">Delete this picture</button>
+</article>
+<mat-divider></mat-divider>
+<article>
+  <form [formGroup]="control" fxLayout="column" fxLayoutAlign="center start">
+    <h4>Informations</h4>
+    <mat-form-field appearance="outline">
+      <mat-label>Address</mat-label>
+      <input matInput placeholder="Position" type="text" formControlName="address" />
+    </mat-form-field>
+    <mat-form-field appearance="outline">
+      <mat-label>Phone number</mat-label>
+      <input matInput placeholder="Phone number" type="tel" formControlName="phoneNumber" />
+    </mat-form-field>
+  </form>
+</article>
