+++ libs/organization/src/lib/components/organization-form/organization-form.component.scss
@@ -1,34 +1,18 @@
-:host {
-  display: block;
-  height: 100%;
-  padding: 2rem;
-  margin: 100px 300px;
-
-  .mat-form-field {
-    width: 100%;
-  }
-
-  .mat-card-title {
-    font-size: 1.8rem;
-  }
-
-  .mat-toolbar {
-    background: transparent;
-
-    .mat-toolbar-row {
-      display: flex;
-      justify-content: center;
-      align-items: center;
-
-      > div {
-        width: 55%;
-      }
-
-      button {
-        margin: 0.5rem;
-        width: 50%;
-      }
-    }
-  }
+mat-form-field {
+  width: 100%;
 }
 
+img {
+  width: 200px;
+  height: 200px;
+}
+
+h4 {
+  margin-bottom: 16px;
+  padding-top: 16px;
+  color: var(--lightGrey);
+}
+
+.actions {
+  padding: 8px 0;
+}
