+++ libs/organization/src/lib/components/organization-form/organization-form.component.ts
@@ -1,12 +1,5 @@
-import { ChangeDetectionStrategy, Component, OnDestroy, OnInit } from '@angular/core';
-import { FormBuilder, FormGroup, Validators } from '@angular/forms';
-import { MatSnackBar } from '@angular/material/snack-bar';
-import { AuthQuery, User } from '@blockframes/auth';
-import { PersistNgFormPlugin } from '@datorama/akita';
-import { first, takeUntil } from 'rxjs/operators';
-import { ActivatedRoute, Router } from '@angular/router';
-import { createOrganization, OrganizationQuery, OrganizationService, OrganizationState } from '../../+state';
-import { Subject } from 'rxjs';
+import { Component, ChangeDetectionStrategy } from '@angular/core';
+import { ControlContainer } from '@angular/forms';
 
 @Component({
   selector: 'organization-form',
@@ -14,71 +7,10 @@ import { Subject } from 'rxjs';
   styleUrls: ['./organization-form.component.scss'],
   changeDetection: ChangeDetectionStrategy.OnPush
 })
-export class OrganizationFormComponent implements OnInit, OnDestroy {
-  public persistForm: PersistNgFormPlugin<OrganizationState>;
-  public user: User;
-  public form: FormGroup;
-  private destroyed$ = new Subject();
+export class OrganizationFormComponent {
+  constructor(public controlContainer: ControlContainer) {}
 
-  constructor(
-    private service: OrganizationService,
-    private query: OrganizationQuery,
-    private auth: AuthQuery,
-    private snackBar: MatSnackBar,
-    private builder: FormBuilder,
-    private route: ActivatedRoute,
-    private router: Router
-  ) {
-  }
-
-  ngOnInit() {
-    this.user = this.auth.user;
-    this.form = this.builder.group({
-      'id': [''],
-      'name': ['', [Validators.required]],
-      'address': ['', [Validators.required]]
-    });
-
-    this.persistForm = new PersistNgFormPlugin(this.query, 'form');
-    this.persistForm.setForm(this.form);
-    this.route.data
-      .pipe(takeUntil(this.destroyed$))
-      .subscribe(({ org }) => org ? this.form.setValue(createOrganization(org)) : this.form.reset());
-  }
-
-  ngOnDestroy() {
-    this.destroyed$.next();
-    this.destroyed$.unsubscribe();
-    this.persistForm.destroy();
-  }
-
-  /////////////
-  // ACTIONS //
-  /////////////
-
-  /** Add a new Organization */
-  public async addOrganization() {
-    if (!this.form.valid) {
-      this.snackBar.open('form invalid', 'close', { duration: 1000 });
-      return;
-    }
-
-    const id = await this.service.add(this.form.value, await this.user);
-
-    this.router.navigate(['layout/o/organization', id])
-    this.snackBar.open(`Created ${this.form.get('name').value}`, 'close', { duration: 1000 });
-    this.form.reset();
-    this.persistForm.reset();
-  }
-
-  /** Clear current form with cancellation */
-  public clear() {
-    const oldState = this.form.value;
-    this.form.reset();
-    this.persistForm.reset();
-    this.snackBar.open('Cleared', 'Cancel', { duration: 1000 })
-      .onAction()
-      .pipe(first())
-      .subscribe(_ => this.form.setValue(oldState));
+  public get control() {
+    return this.controlContainer.control;
   }
 }
