+++ libs/organization/src/lib/pages/organization-editable/organization-editable.component.html
@@ -8,6 +8,6 @@
     </organization-display>
   </ng-container>
   <ng-container sidenav>
-    <organization-edit [formGroup]="organizationProfileForm"></organization-edit>
+    <organization-form [formGroup]="organizationProfileForm"></organization-form>
   </ng-container>
 </editable-sidenav>
